[ClacoForm] fixes categories assignation + cleans ui
