+++ src/plugin/claco-form/Resources/modules/modals/keyword/index.js
@@ -1,12 +1,12 @@
 import {registry} from '#/main/app/modals/registry'
 
 // gets the modal component
-import {KeywordFormModal} from '#/plugin/claco-form/modals/keyword/components/keyword-form'
+import {KeywordModal} from '#/plugin/claco-form/modals/keyword/containers/modal'
 
 const MODAL_KEYWORD_FORM = 'MODAL_KEYWORD_FORM'
 
 // make the modal available for use
-registry.add(MODAL_KEYWORD_FORM, KeywordFormModal)
+registry.add(MODAL_KEYWORD_FORM, KeywordModal)
 
 export {
   MODAL_KEYWORD_FORM
