+++ src/plugin/claco-form/Resources/modules/modals/keyword/store/index.js
@@ -1,0 +1,8 @@
+
+import {reducer} from '#/plugin/claco-form/modals/keyword/store/reducer'
+import {selectors} from '#/plugin/claco-form/modals/keyword/store/selectors'
+
+export {
+  reducer,
+  selectors
+}
