+++ src/plugin/claco-form/Resources/modules/modals/keyword/store/selectors.js
@@ -1,0 +1,6 @@
+
+const STORE_NAME = 'keywordForm'
+
+export const selectors = {
+  STORE_NAME
+}
