+++ src/plugin/claco-form/Resources/modules/resources/claco-form/editor/store/index.js
@@ -4,8 +4,10 @@
 
 import {actions} from '#/plugin/claco-form/resources/claco-form/editor/store/actions'
 import {reducer} from '#/plugin/claco-form/resources/claco-form/editor/store/reducer'
+import {selectors} from '#/plugin/claco-form/resources/claco-form/editor/store/selectors'
 
 export {
   actions,
-  reducer
+  reducer,
+  selectors
 }
