+++ src/plugin/claco-form/Resources/modules/resources/claco-form/editor/store/selectors.js
@@ -1,0 +1,7 @@
+import {selectors as basSelectors} from '#/plugin/claco-form/resources/claco-form/store/selectors'
+
+const STORE_NAME = basSelectors.STORE_NAME+'.clacoFormForm'
+
+export const selectors = {
+  STORE_NAME
+}
