+++ src/plugin/claco-form/Resources/translations/clacoform.en.json
@@ -3,14 +3,11 @@
     "add_choice": "Add a choice",
     "addition": "Addition",
     "all_entries": "All entries",
-    "allowed_file_types": "Allowed file types",
     "associate_category": "Associate a category",
     "audio": "Audio",
     "block": "Block",
     "both": "Both",
     "categories": "Categories",
-    "categories_management": "Categories management",
-    "categories_to_filter": "Categories to filter",
     "change_entry_owner": "Change entry owner",
     "choice_anonymous_comments_only": "Anonymous comments only",
     "choice_entry_all": "All",
@@ -42,14 +39,7 @@
     "delete_selected_entries_confirm_message": "{1} Are you sure you want to delete the selected entry ?|[2,Inf[ Are you sure you want to delete the selected entries ?",
     "display": "Display",
     "display_comments_for_roles": "Display comments for roles",
-    "edit_category": "Edit category",
-    "edit_entry": "Edit entry",
-    "edit_field": "Edit field",
-    "edit_keyword": "Edit keyword",
     "edition": "Edition",
-    "edition_date": "Edition date",
-    "edition_only": "Edition only",
-    "editions": "Editions",
     "enable_comments_for_roles": "Enable comments for roles",
     "entries": "Entries",
     "entries_list": "List of entries",
@@ -75,19 +65,11 @@
     "field_for_title": "Field to display for title",
     "field_for_subtitle": "Field to display for subtitle",
     "fields_associations": "Fields associations",
-    "fields_management": "Fields management",
-    "fields_to_display": "Fields to display",
     "find_entry": "Find an entry",
-    "form_not_blank_error": "This value may not be blank.",
-    "form_not_unique_error": "This value is already used.",
-    "form_number_superior_error": "This value must be %value% or more.",
-    "form_reserved_error": "This value is reserved.",
-    "general_configuration": "General configuration",
-    "hidden_field": "Hidden field",
     "image": "Image",
     "keywords": "Keywords",
-    "keywords_management": "Keywords management",
-    "label_anonymous_comments_enabled": "Allow anonymous comments",
+    "keyword": "Keyword",
+    "new_keyword": "New keyword",
     "label_comments_enabled": "Allow comments",
     "label_creation_enabled": "Allow addition",
     "label_default_home": "Home by default",
@@ -98,7 +80,6 @@
     "label_display_keywords": "Display keywords to users",
     "label_display_metadata": "Display confidential data",
     "label_display_nb_entries": "Display the number of entries",
-    "label_display_votes": "Display votes",
     "label_edition_enabled": "Allow edition",
     "label_keywords_enabled": "Use keywords",
     "label_max_entries": "Number of allowed addition by user (0 = unlimited)",
