+++ src/plugin/claco-form/Resources/translations/clacoform.es.json
@@ -6,7 +6,6 @@
     "associate_category": "",
     "block": "",
     "both": "",
-    "categories_management": "",
     "category_notify_addition": "",
     "category_notify_edition": "",
     "category_notify_removal": "",
@@ -33,11 +32,6 @@
     "delete_keyword": "",
     "delete_keyword_confirm_message": "",
     "display": "Mostrar",
-    "edit_category": "",
-    "edit_field": "",
-    "edit_keyword": "",
-    "edition_date": "",
-    "edition_only": "",
     "entries": "",
     "entries_list": "",
     "entries_list_search": "",
@@ -52,17 +46,8 @@
     "entry_removal_from_category_msg": "",
     "entry_title": "",
     "entry_title_info": "",
-    "fields_management": "",
-    "fields_to_display": "",
     "find_entry": "",
-    "form_not_blank_error": "",
-    "form_not_unique_error": "",
-    "form_number_superior_error": "",
-    "form_reserved_error": "",
-    "general_configuration": "",
     "keywords": "",
-    "keywords_management": "",
-    "label_anonymous_comments_enabled": "",
     "label_comments_enabled": "",
     "label_creation_enabled": "",
     "label_default_home": "",
@@ -73,7 +58,6 @@
     "label_display_keywords": "",
     "label_display_metadata": "",
     "label_display_nb_entries": "",
-    "label_display_votes": "",
     "label_edition_enabled": "",
     "label_keywords_enabled": "",
     "label_max_entries": "",
