+++ src/plugin/claco-form/Resources/translations/clacoform.fr.json
@@ -3,14 +3,11 @@
     "add_choice": "Ajouter un choix",
     "addition": "Ajout",
     "all_entries": "Toutes les fiches",
-    "allowed_file_types": "Types de fichier autorisés",
     "associate_category": "Associer une catégorie",
     "audio": "Audio",
     "block": "Bloquer",
     "both": "Les deux",
     "categories": "Catégories",
-    "categories_management": "Gestion des catégories",
-    "categories_to_filter": "Catégories à filtrer",
     "change_entry_owner": "Changer le propriétaire de la fiche",
     "choice_anonymous_comments_only": "Uniquement les commentaires anonymes",
     "choice_entry_all": "Toutes",
@@ -42,14 +39,7 @@
     "delete_selected_entries_confirm_message": "{1} Êtes-vous sûr de vouloir supprimer la fiche sélectionnée ?|[2,Inf[ Êtes-vous sûr de vouloir supprimer les %count% fiches sélectionnées ?",
     "display": "Afficher",
     "display_comments_for_roles": "Afficher les commentaires pour les roles",
-    "edit_category": "Modifier la catégorie",
-    "edit_entry": "Modifier la fiche",
-    "edit_field": "Modifier le champ",
-    "edit_keyword": "Modifier le mot-clé",
     "edition": "Modification",
-    "edition_date": "Date de modification",
-    "edition_only": "Seulement la modification",
-    "editions": "Modifications",
     "enable_comments_for_roles": "Autoriser les commentaires pour les roles",
     "entries": "Fiches",
     "entries_list": "Liste de fiches",
@@ -75,19 +65,11 @@
     "field_for_title": "Champ à afficher pour le titre",
     "field_for_subtitle": "Champ à afficher pour le sous-titre",
     "fields_associations": "Associations aux champs",
-    "fields_management": "Gestion des champs",
-    "fields_to_display": "Champs à afficher",
     "find_entry": "Trouver une fiche",
-    "form_not_blank_error": "Cette valeur ne doit pas être vide.",
-    "form_not_unique_error": "Cette valeur est déjà utilisée.",
-    "form_number_superior_error": "Cette valeur doit être supérieure ou égale à %value%.",
-    "form_reserved_error": "Cette valeur est réservée.",
-    "general_configuration": "Configuration générale",
-    "hidden_field": "Champ caché",
     "image": "Image",
     "keywords": "Mots-clés",
-    "keywords_management": "Gestion des mots-clés",
-    "label_anonymous_comments_enabled": "Autoriser les commentaires anonymes",
+    "keyword": "Mot-clé",
+    "new_keyword": "Nouveau mot-clé",
     "label_comments_enabled": "Autoriser les commentaires",
     "label_creation_enabled": "Autoriser l'ajout de fiche",
     "label_default_home": "Type d'accueil par défaut",
@@ -98,7 +80,6 @@
     "label_display_keywords": "Afficher les mots-clés aux utilisateurs",
     "label_display_metadata": "Afficher les données confidentielles",
     "label_display_nb_entries": "Indiquer le nombre de fiches",
-    "label_display_votes": "Afficher les votes",
     "label_edition_enabled": "Autoriser l'édition personnelle",
     "label_keywords_enabled": "Utiliser les mots-clés",
     "label_max_entries": "Nombre de depôts autorisés par utilisateur (0 = illimité)",
