+++ src/plugin/claco-form/Resources/translations/clacoform.nl.json
@@ -6,7 +6,6 @@
     "associate_category": "Een categorie associëren",
     "block": "Blokkeren",
     "both": "Beide",
-    "categories_management": "De categorieën beheren",
     "choice_anonymous_comments_only": "Alleen anonieme commentaren",
     "choice_entry_all": "Alle",
     "choice_entry_published": "Gepubliceerd",
@@ -30,12 +29,7 @@
     "delete_keyword": "Sleutelwoord verwijderen",
     "delete_keyword_confirm_message": "Bent u er zeker van dat u dit sleutelwoord [% name%] wenst te verwijderen?",
     "display": "Tonen",
-    "edit_category": "Categorie bewerken",
-    "edit_field": "Veld bewerken",
-    "edit_keyword": "Sleutelwoord bewerken",
-    "edition_date": "Datum bewerken",
     "edition": "Bewerken",
-    "edition_only": "",
     "entries": "Vermeldingen",
     "entries_list": "Lijst van vermeldingen",
     "entries_list_search": "Lijst van vermeldingen (Zoeken)",
@@ -50,17 +44,8 @@
     "entry_removal_from_category_msg": "De vermelding [%title%] van de databank [%clacoform%] is verwijderd van categorie [%category%].",
     "entry_title": "Titel van de vermelding",
     "entry_title_info": "Tekst die gebruikt zal worden als titel van de vermelding.",
-    "fields_management": "De velden beheren",
-    "fields_to_display": "Velden te tonen",
     "find_entry": "Een vermelding vinden",
-    "form_not_blank_error": "Deze waarde mag niet leeg zijn.",
-    "form_not_unique_error": "Deze waarde wordt al gebruikt.",
-    "form_number_superior_error": "Deze waarde moet % value% zijn of meer.",
-    "form_reserved_error": "Deze waarde is gereserveerd.",
-    "general_configuration": "Algemene configuratie",
     "keywords": "Sleutelwoorden",
-    "keywords_management": "De sleutelwoorden beheren",
-    "label_anonymous_comments_enabled": "Anonieme commentaren toelaten",
     "label_comments_enabled": "Commentaren toelaten",
     "label_creation_enabled": "Toevoeging toelaten",
     "label_default_home": "Standaard home",
@@ -71,7 +56,6 @@
     "label_display_keywords": "Toon de sleutelwoorden aan de gebruikers",
     "label_display_metadata": "Vertrouwelijke gegevens tonen",
     "label_display_nb_entries": "Het aantal vermeldingen tonen",
-    "label_display_votes": "De stemmen tonen",
     "label_edition_enabled": "Bewerking toelaten",
     "label_keywords_enabled": "Gebruik sleutelwoorden",
     "label_max_entries": "Aantal toegelaten vermeldingen per gebruiker (0 = onbeperkt)",
