+++ src/plugin/claco-form/Resources/views/claco_form/entries_export.html.twig
@@ -12,7 +12,7 @@
                 <th>{{ 'entry_title'|trans({}, 'clacoform') }}</th>
                 <th>{{ 'author'|trans({}, 'platform') }}</th>
                 <th>{{ 'publication_date'|trans({}, 'clacoform') }}</th>
-                <th>{{ 'edition_date'|trans({}, 'clacoform') }}</th>
+                <th>{{ 'last_modification'|trans({}, 'platform') }}</th>
 
                 {% for field in fields %}
                     <th>{{ field.getName() }}</th>
